# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MemberStatus = (typeof MemberStatus)[keyof typeof MemberStatus];

export const MemberStatus = {
  active: "active",
  expiring_soon: "expiring_soon",
  expired: "expired",
} as const;

export interface Member {
  id: number;
  memberId: string;
  fullName: string;
  phoneNumber: string;
  profilePhotoUrl?: string | null;
  membershipStartDate: string;
  membershipDurationMonths: number;
  membershipEndDate: string;
  status: MemberStatus;
  daysRemaining: number;
  createdAt: string;
}

export interface CreateMemberBody {
  fullName: string;
  phoneNumber: string;
  membershipStartDate: string;
  membershipDurationMonths: number;
  profilePhoto?: Blob;
}

export interface UpdateMemberBody {
  fullName?: string;
  phoneNumber?: string;
  membershipStartDate?: string;
  membershipDurationMonths?: number;
  profilePhoto?: Blob;
}

export interface MemberListResponse {
  members: Member[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type CalendarResponseExpiryDates = { [key: string]: Member[] };

export interface CalendarResponse {
  year: number;
  month: number;
  expiryDates: CalendarResponseExpiryDates;
}

export interface DashboardStats {
  totalMembers: number;
  activeMembers: number;
  expiringSoonMembers: number;
  expiredMembers: number;
  newMembersThisMonth: number;
  renewalsDue: number;
}

export interface NotificationResult {
  sent: number;
  failed: number;
  members: string[];
}

export type NotificationLogStatus =
  (typeof NotificationLogStatus)[keyof typeof NotificationLogStatus];

export const NotificationLogStatus = {
  sent: "sent",
  failed: "failed",
} as const;

export interface NotificationLog {
  id: number;
  memberId: number;
  memberName: string;
  phoneNumber: string;
  message: string;
  status: NotificationLogStatus;
  sentAt: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export type ListMembersParams = {
  search?: string;
  status?: ListMembersStatus;
  page?: number;
  limit?: number;
  sortBy?: string;
  sortOrder?: ListMembersSortOrder;
};

export type ListMembersStatus =
  (typeof ListMembersStatus)[keyof typeof ListMembersStatus];

export const ListMembersStatus = {
  active: "active",
  expiring_soon: "expiring_soon",
  expired: "expired",
  all: "all",
} as const;

export type ListMembersSortOrder =
  (typeof ListMembersSortOrder)[keyof typeof ListMembersSortOrder];

export const ListMembersSortOrder = {
  asc: "asc",
  desc: "desc",
} as const;