# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gym Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: members
    description: Member operations
  - name: dashboard
    description: Dashboard stats
  - name: notifications
    description: Notification operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List members with filtering and pagination
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, expiring_soon, expired, all]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListResponse"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a new member
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateMemberBody"
      responses:
        "201":
          description: Member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a single member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMember
      tags: [members]
      summary: Update a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateMemberBody"
      responses:
        "200":
          description: Member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /members/export/csv:
    get:
      operationId: exportMembersCsv
      tags: [members]
      summary: Export all members as CSV
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

  /members/calendar/{year}/{month}:
    get:
      operationId: getMembersCalendar
      tags: [members]
      summary: Get members expiring in a given month
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Calendar data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/expiring-soon:
    get:
      operationId: getExpiringSoon
      tags: [dashboard]
      summary: Get members expiring within 7 days
      responses:
        "200":
          description: Expiring soon members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"

  /notifications/send:
    post:
      operationId: sendNotifications
      tags: [notifications]
      summary: Manually trigger expiry notifications
      responses:
        "200":
          description: Notifications sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResult"

  /notifications/history:
    get:
      operationId: getNotificationHistory
      tags: [notifications]
      summary: Get notification history
      responses:
        "200":
          description: Notification history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationLog"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Member:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: string
        fullName:
          type: string
        phoneNumber:
          type: string
        profilePhotoUrl:
          type: string
          nullable: true
        membershipStartDate:
          type: string
          format: date
        membershipDurationMonths:
          type: integer
        membershipEndDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, expiring_soon, expired]
        daysRemaining:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - memberId
        - fullName
        - phoneNumber
        - membershipStartDate
        - membershipDurationMonths
        - membershipEndDate
        - status
        - daysRemaining
        - createdAt

    CreateMemberBody:
      type: object
      properties:
        fullName:
          type: string
        phoneNumber:
          type: string
        membershipStartDate:
          type: string
          format: date
        membershipDurationMonths:
          type: integer
        profilePhoto:
          type: string
          format: binary
      required:
        - fullName
        - phoneNumber
        - membershipStartDate
        - membershipDurationMonths

    UpdateMemberBody:
      type: object
      properties:
        fullName:
          type: string
        phoneNumber:
          type: string
        membershipStartDate:
          type: string
          format: date
        membershipDurationMonths:
          type: integer
        profilePhoto:
          type: string
          format: binary

    MemberListResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - members
        - total
        - page
        - limit
        - totalPages

    CalendarResponse:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        expiryDates:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/Member"
      required:
        - year
        - month
        - expiryDates

    DashboardStats:
      type: object
      properties:
        totalMembers:
          type: integer
        activeMembers:
          type: integer
        expiringSoonMembers:
          type: integer
        expiredMembers:
          type: integer
        newMembersThisMonth:
          type: integer
        renewalsDue:
          type: integer
      required:
        - totalMembers
        - activeMembers
        - expiringSoonMembers
        - expiredMembers
        - newMembersThisMonth
        - renewalsDue

    NotificationResult:
      type: object
      properties:
        sent:
          type: integer
        failed:
          type: integer
        members:
          type: array
          items:
            type: string
      required:
        - sent
        - failed
        - members

    NotificationLog:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        phoneNumber:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [sent, failed]
        sentAt:
          type: string
          format: date-time
      required:
        - id
        - memberId
        - memberName
        - phoneNumber
        - message
        - status
        - sentAt

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error