# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym Management API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List members with filtering and pagination
 */
export const listMembersQueryPageDefault = 1;
export const listMembersQueryLimitDefault = 10;

export const ListMembersQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  status: zod.enum(["active", "expiring_soon", "expired", "all"]).optional(),
  page: zod.coerce.number().default(listMembersQueryPageDefault),
  limit: zod.coerce.number().default(listMembersQueryLimitDefault),
  sortBy: zod.coerce.string().optional(),
  sortOrder: zod.enum(["asc", "desc"]).optional(),
});

export const ListMembersResponse = zod.object({
  members: zod.array(
    zod.object({
      id: zod.number(),
      memberId: zod.string(),
      fullName: zod.string(),
      phoneNumber: zod.string(),
      profilePhotoUrl: zod.string().nullish(),
      membershipStartDate: zod.coerce.date(),
      membershipDurationMonths: zod.number(),
      membershipEndDate: zod.coerce.date(),
      status: zod.enum(["active", "expiring_soon", "expired"]),
      daysRemaining: zod.number(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Create a new member
 */
export const CreateMemberBody = zod.object({
  fullName: zod.string(),
  phoneNumber: zod.string(),
  membershipStartDate: zod.coerce.date(),
  membershipDurationMonths: zod.number(),
  profilePhoto: zod.instanceof(File).optional(),
});

/**
 * @summary Get a single member
 */
export const GetMemberParams = zod.object({
  id: zod.coerce.number(),
});

export const GetMemberResponse = zod.object({
  id: zod.number(),
  memberId: zod.string(),
  fullName: zod.string(),
  phoneNumber: zod.string(),
  profilePhotoUrl: zod.string().nullish(),
  membershipStartDate: zod.coerce.date(),
  membershipDurationMonths: zod.number(),
  membershipEndDate: zod.coerce.date(),
  status: zod.enum(["active", "expiring_soon", "expired"]),
  daysRemaining: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update a member
 */
export const UpdateMemberParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateMemberBody = zod.object({
  fullName: zod.string().optional(),
  phoneNumber: zod.string().optional(),
  membershipStartDate: zod.coerce.date().optional(),
  membershipDurationMonths: zod.number().optional(),
  profilePhoto: zod.instanceof(File).optional(),
});

export const UpdateMemberResponse = zod.object({
  id: zod.number(),
  memberId: zod.string(),
  fullName: zod.string(),
  phoneNumber: zod.string(),
  profilePhotoUrl: zod.string().nullish(),
  membershipStartDate: zod.coerce.date(),
  membershipDurationMonths: zod.number(),
  membershipEndDate: zod.coerce.date(),
  status: zod.enum(["active", "expiring_soon", "expired"]),
  daysRemaining: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a member
 */
export const DeleteMemberParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteMemberResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get members expiring in a given month
 */
export const GetMembersCalendarParams = zod.object({
  year: zod.coerce.number(),
  month: zod.coerce.number(),
});

export const GetMembersCalendarResponse = zod.object({
  year: zod.number(),
  month: zod.number(),
  expiryDates: zod.record(
    zod.string(),
    zod.array(
      zod.object({
        id: zod.number(),
        memberId: zod.string(),
        fullName: zod.string(),
        phoneNumber: zod.string(),
        profilePhotoUrl: zod.string().nullish(),
        membershipStartDate: zod.coerce.date(),
        membershipDurationMonths: zod.number(),
        membershipEndDate: zod.coerce.date(),
        status: zod.enum(["active", "expiring_soon", "expired"]),
        daysRemaining: zod.number(),
        createdAt: zod.coerce.date(),
      }),
    ),
  ),
});

/**
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalMembers: zod.number(),
  activeMembers: zod.number(),
  expiringSoonMembers: zod.number(),
  expiredMembers: zod.number(),
  newMembersThisMonth: zod.number(),
  renewalsDue: zod.number(),
});

/**
 * @summary Get members expiring within 7 days
 */
export const GetExpiringSoonResponseItem = zod.object({
  id: zod.number(),
  memberId: zod.string(),
  fullName: zod.string(),
  phoneNumber: zod.string(),
  profilePhotoUrl: zod.string().nullish(),
  membershipStartDate: zod.coerce.date(),
  membershipDurationMonths: zod.number(),
  membershipEndDate: zod.coerce.date(),
  status: zod.enum(["active", "expiring_soon", "expired"]),
  daysRemaining: zod.number(),
  createdAt: zod.coerce.date(),
});
export const GetExpiringSoonResponse = zod.array(GetExpiringSoonResponseItem);

/**
 * @summary Manually trigger expiry notifications
 */
export const SendNotificationsResponse = zod.object({
  sent: zod.number(),
  failed: zod.number(),
  members: zod.array(zod.string()),
});

/**
 * @summary Get notification history
 */
export const GetNotificationHistoryResponseItem = zod.object({
  id: zod.number(),
  memberId: zod.number(),
  memberName: zod.string(),
  phoneNumber: zod.string(),
  message: zod.string(),
  status: zod.enum(["sent", "failed"]),
  sentAt: zod.coerce.date(),
});
export const GetNotificationHistoryResponse = zod.array(
  GetNotificationHistoryResponseItem,
);