# lib/api-zod/src/generated/types/calendarResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarResponseExpiryDates } from "./calendarResponseExpiryDates";

export interface CalendarResponse {
  year: number;
  month: number;
  expiryDates: CalendarResponseExpiryDates;
}