# lib/api-zod/src/generated/types/updateMemberBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym Management API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMemberBody {
  fullName?: string;
  phoneNumber?: string;
  membershipStartDate?: Date;
  membershipDurationMonths?: number;
  profilePhoto?: Blob;
}